+++ maptasker/src/guimap.py
@@ -57,14 +57,15 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
 
     Note: This code is really ugly!
     """
+    lines_to_skip = 0
     # Determine if there is html in current line (typically it is in the line after the current line)
     temp = lines[line_num].split("text-box")
     if "style=" not in temp[1]:
         line_num += 1
+        lines_to_skip += 1
         line_num_to_add = line_num
     else:
         line_num_to_add = line_num + 1
-    lines_to_skip = 0
 
     # A new flag to control the flow of the outer while loop
     continue_processing = True
@@ -86,8 +87,6 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
         lblend = False
 
         for num, line in enumerate(html_lines):
-            if "Build List" in line:
-                print("bingo")
             # Skip empty lines or lines that are just closing span tags
             if not line or line == "</span>" or line.endswith('text-box"><p>'):
                 continue
@@ -166,7 +165,11 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
             if processed_line_data:
                 last_item = processed_line_data[-1]
                 # Compare the current color and font with the last one
-                if last_item["color"] == color and last_item["highlights"] == font:
+                if (
+                    last_item["color"] == color
+                    and last_item["highlights"] == font
+                    and not ("<a href=" in text or "<a href=" in last_item["text"])
+                ):
                     # If they match, concatenate the text in the processed_line_data list
                     last_item["text"] += text
                     last_item["end"] = lblend
@@ -175,7 +178,15 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
                     if not lblend:
                         continue
                 elif lblend:
-                    processed_line_data = add_line_data(processed_line_data, text, color, font, spacing, lblend, decor)
+                    processed_line_data = add_line_data(
+                        processed_line_data,
+                        text,
+                        color,
+                        font,
+                        spacing,
+                        lblend,
+                        decor,
+                    )
                     break
 
                 # Reset spacing since we're now adding to an existing line.
@@ -187,7 +198,15 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
 
             # If they don't match or it's the first element or the color/font don't match previous...
             # add a new entry to processed_line_data
-            processed_line_data = add_line_data(processed_line_data, text, color, font, spacing, lblend, decor)
+            processed_line_data = add_line_data(
+                processed_line_data,
+                text,
+                color,
+                font,
+                spacing,
+                lblend,
+                decor,
+            )
 
         line_num += 1
         lines_to_skip += 1
@@ -934,8 +953,13 @@ def process_html_lines(
     lines_to_skip = 0
 
     for line_num, line in enumerate(lines):
-        if "Build List" in line:
+        if "OUTPUT TEMPLATE" in line:
             print("bingo")
+        # Handle labels and TaskerNet descriptions with html in them.
+        # FIX lines-to-skip is foobar'ed
+        if "text-box" in line and ".text-box" not in line:
+            lines_to_skip = process_label_html(lines, output_lines, line_num, spacing)
+            # continue
         # Are we to skip lines due to label with html having already been added?
         if lines_to_skip > 0:
             lines_to_skip -= 1
@@ -1002,10 +1026,10 @@ def process_html_lines(
             remove_html,
         )
 
-        # Handle labels and TaskerNet descriptions with html in them.
-        if "text-box" in line and ".text-box" not in line:
-            lines_to_skip = process_label_html(lines, output_lines, line_num, spacing)
-            continue
+        # # Handle labels and TaskerNet descriptions with html in them.
+        # if "text-box" in line and ".text-box" not in line:
+        #     lines_to_skip = process_label_html(lines, output_lines, line_num, spacing)
+        #     continue
 
         # If at end of valid html, start removing html again
         if "/html" in line or "<div <span" in line: